cleaned up spec_helper and rails_helper
