+++ .rspec
@@ -1,2 +1,2 @@
 --color
---require spec_helper
+--require rails_helper
