+++ spec/controllers/lesson_completions_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe LessonCompletionsController do
