+++ spec/controllers/omniauth_callbacks_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe OmniauthCallbacksController, "github callback" do
   before do
