+++ spec/controllers/users_controller_spec.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require "rails_helper"
 
 describe UsersController do
