+++ spec/features/auth_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe "Authentication" do
 
@@ -89,11 +89,3 @@ describe "Authentication" do
     end
   end
 end
-
-
-
-
-
-
-
-
