+++ spec/features/smart_redirect.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe "Smart Redirect" do
     #create a course with content
