+++ spec/helpers/lessons_helper_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe LessonsHelper do
