+++ spec/mailers/user_mailer_spec.rb
@@ -1,5 +1,5 @@
-require "spec_helper"
+require "rails_helper"
 
 describe UserMailer do
-  
+
 end
