+++ spec/models/lesson_completion_spec.rb
@@ -1,11 +1,11 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe LessonCompletion do
-  
+
   let!(:user){ FactoryGirl.create(:user) }
   let!(:lesson){ FactoryGirl.create(:lesson) }
-  
-  subject(:lesson_completion) { LessonCompletion.new(   
+
+  subject(:lesson_completion) { LessonCompletion.new(
     :student_id => user.id,
     :lesson_id => lesson.id
                                )}
@@ -20,7 +20,7 @@ describe LessonCompletion do
     end
     it { is_expected.not_to be_valid }
   end
-  
+
   # Also test that they delete when the lesson or user are deleted?
 
 end
