+++ spec/models/user_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 # currently allows overlapping events!!
