+++ spec/tasks/curriculum_spec.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require "rails_helper"
 require "rake"
 
 describe ":curriculum" do
